requirements.txt
#
# This file is autogenerated by pip-compile with Python 3.9
# by the following command:
#
#    pip-compile
#
asgiref==3.6.0
    # via django
django==4.1.5
    # via
    #   -r requirements.in
    #   djangorestframework
djangorestframework==3.14.0
    # via -r requirements.in
pytz==2022.7.1
    # via djangorestframework
sqlparse==0.4.3
    # via django
